requirements.txt
pymongo
requests
bs4
csv
mysql-connector-python